requirements.txt
aenum==3.1.15
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dataclasses-json==0.6.4
fiona==1.9.5
geopandas==0.14.3
greenlet==3.0.3
idna==3.6
marshmallow==3.20.2
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pymssql==2.2.11
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sentinelhub==3.10.1
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.25
tifffile==2024.1.30
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
utm==0.7.0
